Revert "merge upstream changes before adding support for torchbench (#8)" (#10)
This reverts commit fd3a6cd4cfa23cfbdad33f6b5b91739bbc61e47d.
