+++ .ci/docker/requirements.txt
@@ -1,5 +1,4 @@
 torch >= 2.3.0
-torchdata >= 0.8.0
 datasets >= 2.19.0
 tomli >= 1.1.0 ; python_version < "3.11"
 tensorboard
