+++ test/datasets/test_checkpoint.py
@@ -6,13 +6,13 @@
 
 import torch
 from torchtitan.datasets.hf_datasets import build_hf_data_loader
-from torchtitan.datasets.tokenizer import build_tokenizer
+from torchtitan.datasets.tokenizer import create_tokenizer
 
 
 class TestCheckpoint:
     def test_c4_resumption(self):
-        dataset_name = "c4_test"
-        dataset_path = "./test/assets/c4_test"
+        dataset_name = "c4_mini"
+        dataset_path = "./torchtitan/datasets/c4_mini"
         batch_size = 1
         seq_len = 1024
         world_size = 4
@@ -42,7 +42,7 @@ class TestCheckpoint:
         self, dataset_name, dataset_path, batch_size, seq_len, world_size, rank
     ):
         tokenizer_type = "tiktoken"
-        tokenizer = build_tokenizer("tiktoken", "./test/assets/test_tiktoken.model")
+        tokenizer = create_tokenizer("tiktoken", "./test/assets/test_tiktoken.model")
         return build_hf_data_loader(
             dataset_name=dataset_name,
             dataset_path=dataset_path,
