+++ test_runner.py
@@ -50,21 +50,6 @@ def build_test_list():
             [
                 [
                     "--checkpoint.enable_checkpoint",
-                    "--experimental.pipeline_parallel_degree 4",
-                    "--experimental.pipeline_parallel_split_points layers.1,layers.2,layers.3,layers.4,layers.5,layers.6,layers.7",
-                    "--experimental.pipeline_parallel_schedule flexible_interleaved_1f1b",
-                    "--model.norm_type rmsnorm",  # fused_rmsnorm throws cuda context error with pp
-                ],
-            ],
-            "PP looped flexible 1f1b test",
-            "pp_looped_flexible_1f1b",
-            requires_seed_checkpoint=True,
-            ngpu=4,
-        ),
-        OverrideDefinitions(
-            [
-                [
-                    "--checkpoint.enable_checkpoint",
                     "--experimental.pipeline_parallel_degree 2",
                     "--experimental.pipeline_parallel_split_points layers.4",
                     "--experimental.pipeline_parallel_schedule 1f1b",
@@ -167,17 +152,6 @@ def build_test_list():
             ],
             "1D compile",
             "1d_compile",
-        ),
-        OverrideDefinitions(
-            [
-                [
-                    "--training.compile",
-                    "--activation_checkpoint.mode selective",
-                    "--activation_checkpoint.selective_ac_option op",
-                ],
-            ],
-            "1D compile with selective op AC",
-            "1d_compile_sac_op",
         ),
         OverrideDefinitions(
             [
@@ -299,16 +273,6 @@ def build_test_list():
             "fsdp2_mem_tracker",
             ngpu=4,
         ),
-        OverrideDefinitions(
-            [
-                [
-                    "--training.data_parallel_type ddp",
-                ]
-            ],
-            "DDP",
-            "ddp",
-            ngpu=4,
-        ),
     ]
     return integration_tests_flavors
 
@@ -340,8 +304,6 @@ def run_test(test_flavor: OverrideDefinitions, full_path: str, output_dir: str):
 
     for override_arg in test_flavor.override_args:
         cmd = f"CONFIG_FILE={full_path} NGPU={test_flavor.ngpu} LOG_RANK={all_ranks} ./run_llama_train.sh"
-        if test_name == "fsdp2_mem_tracker":
-            cmd = f"CONFIG_FILE={full_path} NGPU={test_flavor.ngpu} LOG_RANK={all_ranks} ./run_memory_estimation.sh"
         cmd += " " + dump_folder_arg
         cmd += " " + model_flavor_arg
         if override_arg:
