+++ torchtitan/datasets/tokenizer/sentencepiece.py
@@ -11,7 +11,7 @@ from typing import List
 from sentencepiece import SentencePieceProcessor
 
 from torchtitan.datasets.tokenizer.tokenizer import Tokenizer
-from torchtitan.logging import logger
+from torchtitan.logging_utils import logger
 
 
 class SentencePieceTokenizer(Tokenizer):
