+++ torchtitan/models/llama/__init__.py
@@ -48,13 +48,4 @@ llama3_configs = {
         multiple_of=4096,
         rope_theta=500000,
     ),
-    "405B": ModelArgs(
-        dim=16384,
-        n_layers=126,
-        n_heads=128,
-        n_kv_heads=8,
-        ffn_dim_multiplier=1.2,
-        multiple_of=4096,
-        rope_theta=500000,
-    ),
 }
