+++ torchtitan/models/norms.py
@@ -18,18 +18,18 @@ from torch.distributed._tensor import Partial, Replicate, Shard
 from torch.distributed._tensor.experimental import local_map
 
 
-def build_norm(norm_type: str, dim: int, eps: float = 1e-6):
+def create_norm(norm_type: str, dim: int, eps: float = 1e-6):
     """
-    Builds the specified normalization layer based on the norm_type.
+    Creates the specified normalization layer based on the norm_type.
 
     Args:
-        norm_type (str): The type of normalization layer to build.
+        norm_type (str): The type of normalization layer to create.
             Supported types: 1. rmsnorm 2. fused_rmsnorm 3. layernorm 4. np_layernorm
         dim (int): The dimension of the normalization layer.
         eps (float, optional): The epsilon value for numerical stability. Defaults to 1e-6.
 
     Returns:
-        The built normalization layer.
+        The created normalization layer.
 
     Raises:
         NotImplementedError: If an unknown norm_type is provided.
@@ -43,11 +43,6 @@ def build_norm(norm_type: str, dim: int, eps: float = 1e-6):
     elif norm_type == "rmsnorm":
         return RMSNorm(dim, eps=eps)
     elif norm_type == "compiled_rmsnorm":
-        import warnings
-
-        warnings.warn(
-            "compiled_rmsnorm is currently experimental and not ready to use yet."
-        )
         return RMSNorm(dim, eps=eps, compile=True)
     elif norm_type == "fused_rmsnorm":
         return FusedRMSNorm(dim, eps=eps)
