+++ torchtitan/parallelisms/parallelize_llama.py
@@ -4,20 +4,23 @@
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
 
-# This file applies the PT-D parallelisms (except pipeline parallelism) and various
-# training techniques (e.g. activation checkpointing and compile) to the Llama model.
+# This file applies the PT-D parallelisms and various training techniques (e.g.
+# activation checkpointing and compile) to the Llama model.
 
+import copy
 from collections import defaultdict
+from typing import Tuple, TYPE_CHECKING, Union
 
 import torch
 import torch.nn as nn
 from torch.distributed import DeviceMesh
+
 from torch.distributed._composable.fsdp import fully_shard, MixedPrecisionPolicy
-from torch.distributed._composable.replicate import replicate
 from torch.distributed._tensor import Replicate, Shard
 from torch.distributed.algorithms._checkpoint.checkpoint_wrapper import (
     checkpoint_wrapper as ptd_checkpoint_wrapper,
 )
+from torch.distributed.pipelining import pipeline, PipelineStage, SplitPoint
 from torch.distributed.tensor.parallel import (
     ColwiseParallel,
     parallelize_module,
@@ -27,219 +30,18 @@ from torch.distributed.tensor.parallel import (
 )
 
 from torchtitan.config_manager import JobConfig, TORCH_DTYPE_MAP
-from torchtitan.logging import logger
-from torchtitan.parallelisms.parallel_dims import ParallelDims
+from torchtitan.logging_utils import logger
+from torchtitan.models.llama.model import ModelArgs
+from torchtitan.parallelisms.pipelining_utils import stage_ids_this_rank
+
+if TYPE_CHECKING:
+    from torchtitan.parallelisms import ParallelDims
 
 
-# NOTE(lty): experimental for the PT-D 24 research internship project
-def parallelize_llama_torch_spmd(
-    model, world_mesh, parallel_dims, job_config: JobConfig
-):
-    assert not parallel_dims.pp_enabled, "PP not supported by torch_spmd yet"
-    assert not parallel_dims.tp_enabled, "TP not supported by torch_spmd yet"
-    ac_config = job_config.activation_checkpoint
-    assert ac_config.mode == "none", "AC not supported by torch_spmd yet"
+DeviceType = Union[int, str, torch.device]
 
-    if parallel_dims.dp_enabled:
-        from torch_spmd.data_parallel import data_parallel, MixedPrecisionPolicy
-
-        torch._inductor.config.simplefsdp.enable_bucket = False
-
-        mp_policy = MixedPrecisionPolicy(
-            param_dtype=TORCH_DTYPE_MAP[job_config.training.mixed_precision_param],
-            reduce_dtype=TORCH_DTYPE_MAP[job_config.training.mixed_precision_reduce],
-        )
-        dp_mesh = world_mesh["dp"] if world_mesh.ndim > 1 else world_mesh
-        model = data_parallel(model, dp_mesh, mode="fully_shard", mp_policy=mp_policy)
-        logger.info("Applied Simple FSDP to the model")
-
-    if job_config.training.compile:
-        torch._dynamo.config.inline_inbuilt_nn_modules = True
-        logger.info("Compiling with torch.compile")
-        model = torch.compile(model, fullgraph=True)
-
-    return model
-
-
-def parallelize_llama(
-    model: nn.Module,
-    world_mesh: DeviceMesh,
-    parallel_dims: ParallelDims,
-    job_config: JobConfig,
-):
-    """
-    Apply tensor parallelism, activation checkpointing, torch.compile, and data
-    parallelism to the model.
-
-    NOTE: The passed-in model preferably should be on meta device. Otherwise,
-    the model must fit on GPU or CPU memory.
-    """
-    # NOTE(lty): experimental for the PT-D 24 research internship project
-    if job_config.experimental.torch_spmd:
-        return parallelize_llama_torch_spmd(
-            model, world_mesh, parallel_dims, job_config
-        )
-
-    if parallel_dims.tp_enabled:
-        if (
-            job_config.experimental.enable_async_tensor_parallel
-            and not job_config.training.compile
-        ):
-            raise RuntimeError("Async TP requires --training.compile")
-        apply_tp(
-            model,
-            world_mesh["tp"],
-            loss_parallel=parallel_dims.loss_parallel_enabled,
-            enable_float8=job_config.float8.enable_float8_linear,
-            enable_async_tp=job_config.experimental.enable_async_tensor_parallel,
-        )
-
-    if job_config.activation_checkpoint.mode != "none":
-        apply_ac(model, job_config.activation_checkpoint)
-
-    # # turn on per-TransformerBlock compile after AC wrapping and before FSDP
-    # if job_config.training.compile:
-    #     if job_config.model.norm_type == "fused_rmsnorm":
-    #         raise NotImplementedError(
-    #             "fused_rmsnorm is not compatible with torch.compile yet. "
-    #             "Please use rmsnorm or layernorm."
-    #         )
-    #     apply_compile(model)
-
-    if parallel_dims.dp_enabled:
-        if parallel_dims.dp_type == "fsdp":
-            dp_mesh = world_mesh["dp"] if world_mesh.ndim > 1 else world_mesh
-            assert dp_mesh.mesh_dim_names == ("dp",), dp_mesh.mesh_dim_names
-
-            apply_fsdp(
-                model,
-                dp_mesh,
-                param_dtype=TORCH_DTYPE_MAP[job_config.training.mixed_precision_param],
-                reduce_dtype=TORCH_DTYPE_MAP[
-                    job_config.training.mixed_precision_reduce
-                ],
-                pp_enabled=parallel_dims.pp_enabled,
-            )
-        else:
-            if world_mesh.ndim > 1:
-                raise RuntimeError("DDP has not supported > 1D parallelism")
-            apply_ddp(
-                model,
-                world_mesh,
-                enable_compile=job_config.training.compile,
-                enable_compiled_autograd=job_config.experimental.enable_compiled_autograd,
-            )
-
-    # turn on whole-model compile after FSDP
-    if job_config.training.compile:
-        if job_config.model.norm_type == "fused_rmsnorm":
-            raise NotImplementedError(
-                "fused_rmsnorm is not compatible with torch.compile yet. "
-                "Please use rmsnorm or layernorm."
-            )
-        model = apply_compile(model)
-
-    return model
-
-
-def apply_tp(
-    model: nn.Module,
-    tp_mesh: DeviceMesh,
-    loss_parallel: bool,
-    enable_float8: bool,
-    enable_async_tp: bool,
-):
-    """Apply tensor parallelism."""
-    # 1. Parallelize the embedding and shard its outputs (which are the first
-    # transformer block's inputs)
-    # 2. Parallelize the root norm layer over the sequence dim
-    # 3. Parallelize the final linear output layer
-    parallelize_module(
-        model,
-        tp_mesh,
-        {
-            "tok_embeddings": RowwiseParallel(
-                input_layouts=Replicate(),
-                output_layouts=Shard(1),
-            ),
-            "norm": SequenceParallel(),
-            "output": ColwiseParallel(
-                input_layouts=Shard(1),
-                output_layouts=Shard(-1) if loss_parallel else Replicate(),
-                use_local_output=not loss_parallel,
-            ),
-        },
-    )
-
-    # Parallel styles used for transformer block linear weights and their
-    # inputs may be different for float8 linears
-    if enable_float8:
-        # TODO(vkuzo): once float8 configuration supports delayed scaling,
-        # add a check here to enforce supported float8 all-gather configurations
-        # TODO(vkuzo): add the items below to __init__.py of torchao.float8 and import from there
-        from torchao.float8.float8_tensor_parallel import (
-            Float8ColwiseParallel,
-            Float8RowwiseParallel,
-            PrepareFloat8ModuleInput,
-        )
-
-        rowwise_parallel, colwise_parallel, prepare_module_input = (
-            Float8RowwiseParallel,
-            Float8ColwiseParallel,
-            PrepareFloat8ModuleInput,
-        )
-    else:
-        rowwise_parallel, colwise_parallel, prepare_module_input = (
-            RowwiseParallel,
-            ColwiseParallel,
-            PrepareModuleInput,
-        )
-
-    # Apply tensor + sequence parallelism to every transformer block
-    # NOTE: At the cost of model code change, we can accelerate Sequence Parallel
-    #       by folding (and unfolding) the batch dimension and the sequence dimension.
-    #       Examples can be found at https://github.com/pytorch/torchtitan/pull/437
-    for layer_id, transformer_block in model.layers.items():
-        layer_plan = {
-            "attention_norm": SequenceParallel(),
-            "attention": prepare_module_input(
-                input_layouts=(Shard(1), None),
-                desired_input_layouts=(Replicate(), None),
-            ),
-            "attention.wq": colwise_parallel(),
-            "attention.wk": colwise_parallel(),
-            "attention.wv": colwise_parallel(),
-            "attention.wo": rowwise_parallel(output_layouts=Shard(1)),
-            "ffn_norm": SequenceParallel(),
-            "feed_forward": prepare_module_input(
-                input_layouts=(Shard(1),),
-                desired_input_layouts=(Replicate(),),
-            ),
-            "feed_forward.w1": colwise_parallel(),
-            "feed_forward.w2": rowwise_parallel(output_layouts=Shard(1)),
-            "feed_forward.w3": colwise_parallel(),
-        }
-
-        parallelize_module(
-            module=transformer_block,
-            device_mesh=tp_mesh,
-            parallelize_plan=layer_plan,
-        )
-
-    if enable_async_tp:
-        from torch.distributed._symmetric_memory import enable_symm_mem_for_group
-
-        torch._inductor.config._micro_pipeline_tp = True
-        enable_symm_mem_for_group(tp_mesh.get_group().group_name)
-
-    logger.info(
-        f"Applied {'Float8 ' if enable_float8 else ''}{'Async ' if enable_async_tp else ''}"
-        "Tensor Parallelism to the model"
-    )
-
-
-# for selective op activation checkpointing
-_save_list = {
+# for selective AC
+no_recompute_list = {
     torch.ops.aten.mm.default,
     torch.ops.aten._scaled_dot_product_efficient_attention.default,
     torch.ops.aten._scaled_dot_product_flash_attention.default,
@@ -247,7 +49,7 @@ _save_list = {
 }
 
 
-def _apply_ac_to_transformer_block(module: nn.Module, ac_config):
+def checkpoint_wrapper(module: torch.nn.Module, ac_config):
     valid_ac_modes = ("full", "selective")
     if ac_config.mode not in valid_ac_modes:
         raise ValueError(
@@ -278,7 +80,7 @@ def _apply_ac_to_transformer_block(module: nn.Module, ac_config):
                 if func == torch.ops.aten.mm.default:
                     meta[mm_count_key] += 1
                 # Saves output of all compute ops, except every second mm
-                to_save = func in _save_list and not (
+                to_save = func in no_recompute_list and not (
                     func == torch.ops.aten.mm.default and meta[mm_count_key] % 2 == 0
                 )
                 return (
@@ -301,6 +103,10 @@ def _apply_ac_to_transformer_block(module: nn.Module, ac_config):
     elif use_layer_sac:
         # Checkpoint every `ac_freq` of the modules passed to this function
         ac_freq = int(ac_config.selective_ac_option)
+        if ac_freq <= 0:
+            raise ValueError(
+                f"Selective layer AC expects a positive int as selective_ac_option but got {ac_freq}"
+            )
         ptd_checkpoint_wrapper.__dict__.setdefault("_count", 0)
         ptd_checkpoint_wrapper._count += 1
         if not ac_freq or ptd_checkpoint_wrapper._count % ac_freq == 0:
@@ -309,25 +115,331 @@ def _apply_ac_to_transformer_block(module: nn.Module, ac_config):
             return module
 
 
-def apply_ac(model: nn.Module, ac_config):
+def get_tp_parallel_strategy(
+    job_config: JobConfig,
+) -> Tuple[RowwiseParallel, ColwiseParallel, PrepareModuleInput]:
+    """Get the parallel strategy for the transformer model.
+
+    This function handles the special case of using float8 with tensor parallelism.
+    """
+    if job_config.training.fp8_linear == "dynamic":
+        from float8_experimental.float8_tensor_parallel import (
+            Float8ColwiseParallel,
+            Float8RowwiseParallel,
+            PrepareFloat8ModuleInput,
+        )
+
+        return Float8RowwiseParallel, Float8ColwiseParallel, PrepareFloat8ModuleInput
+    return RowwiseParallel, ColwiseParallel, PrepareModuleInput
+
+
+def pipeline_llama(
+    model: nn.Module,
+    world_mesh: DeviceMesh,
+    parallel_dims: "ParallelDims",
+    job_config: JobConfig,
+    device: DeviceType,
+    model_config: ModelArgs,
+):
+    split_mode = job_config.experimental.pipeline_parallel_split_mode
+    valid_split_modes = ("manual", "tracer")
+    if split_mode not in valid_split_modes:
+        raise ValueError(
+            f"Invalid split mode: {split_mode}. Valid split modes: {valid_split_modes}"
+        )
+    if split_mode == "manual":
+        return pipeline_llama_manual(
+            model, world_mesh, parallel_dims, job_config, device, model_config
+        )
+    elif split_mode == "tracer":
+        return pipeline_llama_tracer(
+            model, world_mesh, parallel_dims, job_config, device, model_config
+        )
+
+
+def _llama_trace_input(job_config: JobConfig, model_config: ModelArgs, device="meta"):
+    """Get meta tensors with the right input shapes used for tracing"""
+    tokens_shape = (job_config.training.batch_size, job_config.training.seq_len)
+    tokens = torch.randint(
+        model_config.vocab_size, tokens_shape, dtype=torch.int64, device=device
+    )
+    return (tokens,)
+
+
+def _mixed_precision_dtype(
+    job_config: JobConfig, parallel_dims, default: torch.dtype = torch.float32
+) -> torch.dtype:
+    """Get the mixed precision dtype if FSDP is enabled, otherwise return the default"""
+    mp_arg = job_config.training.mixed_precision_param
+    return TORCH_DTYPE_MAP[mp_arg] if parallel_dims.dp_enabled else default
+
+
+def pipeline_llama_manual(
+    whole_model: nn.Module,
+    world_mesh: DeviceMesh,
+    parallel_dims: "ParallelDims",
+    job_config: JobConfig,
+    device: DeviceType,
+    model_config: ModelArgs,
+):
+    """
+    This API extracts one torch.nn.Module objects for the part of the model configured to run inside this stage.
+
+    It wraps the model chunk in a ManualPipelineStage object and returns both the stage and model objects.
+
+    The stage object is used to create a pipeline schedule, and the model object can be used for applying SPMD
+    parallelism.
+    """
+    pp_mesh = world_mesh["pp"]
+    pp_rank = pp_mesh.get_local_rank()
+    pp_size = pp_mesh.size()
+    microbatches = (
+        job_config.experimental.pipeline_parallel_microbatches or parallel_dims.pp
+    )
+    splits = job_config.experimental.pipeline_parallel_split_points
+
+    def _build_stage(stage_idx, start_layer, stop_layer, is_first=False, is_last=False):
+        model = copy.deepcopy(whole_model)
+        if not is_first:
+            model.tok_embeddings = None
+
+        drop_layers = start_layer is not None
+        for name in list(model.layers.keys()):
+            # we keep layers in a contiguous region between start (inclusive) and stop (exclusive)
+            if f"layers.{name}" == start_layer:
+                drop_layers = False
+            if f"layers.{name}" == stop_layer:
+                drop_layers = True
+            if drop_layers:
+                del model.layers[name]
+
+        if not is_last:
+            model.norm = None
+            model.output = None
+
+        # TODO(whc) once ManualPipelineStage supports lazy shape inference, we can leave model on meta device longer and
+        # get rid of the input shape hardcoded here. For now, it should not be a big deal since we only materialize the
+        # layers of the model that map to this stage, not the whole model.
+        mp_dtype = _mixed_precision_dtype(job_config, parallel_dims)
+        batch_size = job_config.training.batch_size
+        local_seq_len = int(job_config.training.seq_len // parallel_dims.tp)
+        layers_io_shape = (batch_size, local_seq_len, model_config.dim)
+        output_layer_shape = (
+            batch_size,
+            job_config.training.seq_len,
+            model_config.vocab_size,
+        )
+        if is_first:
+            (input,) = _llama_trace_input(job_config, model_config, device=device)
+        else:
+            # later layers (assume all start w/ a transformer layer)
+            input = torch.rand(layers_io_shape, dtype=mp_dtype, device=device)
+
+        if is_last:
+            output = torch.rand(output_layer_shape, dtype=torch.float32, device=device)
+        else:
+            # earlier layers (assume all end in a transformer layer)
+            output = torch.rand(layers_io_shape, dtype=mp_dtype, device=device)
+
+        model.to_empty(device=device)
+        stage = PipelineStage(
+            model,
+            stage_idx,
+            num_stages,
+            device,
+            input_args=input.chunk(microbatches)[0],
+            output_args=output.chunk(microbatches)[0],
+            group=pp_mesh.get_group("pp"),
+        )
+        return stage, model
+
+    num_stages = len(splits) + 1
+    stage_idx = pp_rank
+
+    stages = []
+    models = []
+    for stage_idx in stage_ids_this_rank(pp_rank, pp_size, num_stages, style="loop"):
+        start_layer = splits[stage_idx - 1] if stage_idx > 0 else None
+        stop_layer = splits[stage_idx] if stage_idx < num_stages - 1 else None
+        stage, model_chunk = _build_stage(
+            stage_idx,
+            start_layer,
+            stop_layer,
+            is_first=stage_idx == 0,
+            is_last=stage_idx == num_stages - 1,
+        )
+        logger.info(
+            f"PP rank {pp_rank} is building stage_idx {stage_idx}"
+            f" with start_layer {start_layer}, stop_layer {stop_layer}: model chunk \n{model_chunk}"
+        )
+        stages.append(stage)
+        models.append(model_chunk)
+    return stages, models
+
+
+def pipeline_llama_tracer(
+    model: nn.Module,
+    world_mesh: DeviceMesh,
+    parallel_dims: "ParallelDims",
+    job_config: JobConfig,
+    device: DeviceType,
+    model_config: ModelArgs,
+):
+    if job_config.model.norm_type == "fused_rmsnorm":
+        # TODO(whc) - torch._dynamo.exc.Unsupported: Illegal getattr
+        # invocation stride in strict mode from `if dy.stride(-1) != 1:` in
+        # fused_rmsnorm
+        raise NotImplementedError(
+            "fused_rmsnorm is not compatible with Pipeline Tracer yet. Please use rmsnorm or layernorm."
+        )
+    if _mixed_precision_dtype(job_config, parallel_dims) != torch.float32:
+        raise NotImplementedError(
+            "Pipeline tracer does not work with FSDP mixed precision yet. "
+            "To work around, set mixed_precision_param to float32."
+        )
+
+    pp_mesh = world_mesh["pp"]
+    pp_rank = pp_mesh.get_local_rank()
+    pp_size = pp_mesh.size()
+    microbatches = (
+        job_config.experimental.pipeline_parallel_microbatches or parallel_dims.pp
+    )
+    (input,) = _llama_trace_input(job_config, model_config, device=device)
+    stage_idx = pp_rank
+    split_spec = {
+        layer_name: SplitPoint.BEGINNING
+        for layer_name in job_config.experimental.pipeline_parallel_split_points
+    }
+    num_stages = len(split_spec) + 1
+    pipe = pipeline(
+        model,
+        mb_args=(input.chunk(microbatches)[0],),
+        split_spec=split_spec,
+    )
+
+    stages = []
+    models = []
+    for stage_idx in stage_ids_this_rank(pp_rank, pp_size, num_stages, style="loop"):
+        models.append(pipe.get_stage_module(stage_idx))
+        stages.append(
+            pipe.build_stage(
+                stage_idx,
+                device=device,
+                group=pp_mesh.get_group(),
+            )
+        )
+    return (stages, models)
+
+
+def apply_tp(
+    model: nn.Module,
+    world_mesh: DeviceMesh,
+    parallel_dims: "ParallelDims",
+    job_config: JobConfig,
+):
+    """Apply tensor parallelism."""
+
+    tp_mesh = world_mesh["tp"]
+    # Parallel styles used for transformer block linear weights and their
+    # inputs may be different for float8 linears
+    (
+        rowwise_parallel_weight,
+        colwise_parallel_weight,
+        prepare_module_input,
+    ) = get_tp_parallel_strategy(job_config)
+    loss_parallel = parallel_dims.loss_parallel_enabled
+
+    # 1. Parallelize the embedding and shard its outputs (which are the first
+    # transformer block's inputs)
+    # 2. Parallelize the root norm layer over the sequence dim
+    # 3. Parallelize the final linear output layer
+    model = parallelize_module(
+        model,
+        tp_mesh,
+        {
+            "tok_embeddings": RowwiseParallel(
+                input_layouts=Replicate(),
+                output_layouts=Shard(1),
+            ),
+            "norm": SequenceParallel(),
+            "output": colwise_parallel_weight(
+                input_layouts=Shard(1),
+                output_layouts=Shard(-1) if loss_parallel else Replicate(),
+                use_local_output=not loss_parallel,
+            ),
+        },
+    )
+
+    # Apply tensor + sequence parallelism to every transformer block
+    # NOTE: At the cost of model code change, we can accelerate Sequence Parallel
+    #       by folding (and unfolding) the batch dimension and the sequence dimension.
+    #       Examples can be found at https://github.com/pytorch/torchtitan/pull/437
+    for layer_id, transformer_block in model.layers.items():
+        layer_plan = {
+            "attention_norm": SequenceParallel(),
+            "attention": prepare_module_input(
+                input_layouts=(Shard(1), None),
+                desired_input_layouts=(Replicate(), None),
+            ),
+            "attention.wq": colwise_parallel_weight(),
+            "attention.wk": colwise_parallel_weight(),
+            "attention.wv": colwise_parallel_weight(),
+            "attention.wo": rowwise_parallel_weight(output_layouts=Shard(1)),
+            "ffn_norm": SequenceParallel(),
+            "feed_forward": prepare_module_input(
+                input_layouts=(Shard(1),),
+                desired_input_layouts=(Replicate(),),
+            ),
+            "feed_forward.w1": colwise_parallel_weight(),
+            "feed_forward.w2": rowwise_parallel_weight(output_layouts=Shard(1)),
+            "feed_forward.w3": colwise_parallel_weight(),
+        }
+
+        # Adjust attention module to use the local number of heads
+        attn_layer = transformer_block.attention
+        attn_layer.n_heads = attn_layer.n_heads // tp_mesh.size()
+        attn_layer.n_kv_heads = attn_layer.n_kv_heads // tp_mesh.size()
+
+        parallelize_module(
+            module=transformer_block,
+            device_mesh=tp_mesh,
+            parallelize_plan=layer_plan,
+        )
+
+    if job_config.experimental.enable_async_tensor_parallel:
+        from torch.distributed._symmetric_memory import enable_symm_mem_for_group
+
+        torch._inductor.config._micro_pipeline_tp = True
+        enable_symm_mem_for_group(tp_mesh.get_group().group_name)
+
+    logger.info("Applied Tensor Parallelism to the model")
+    return model
+
+
+def apply_ac(model: nn.Module, job_config: JobConfig):
     """Apply activation checkpointing to the model."""
+
+    ac_config = job_config.activation_checkpoint
+
     for layer_id, transformer_block in model.layers.named_children():
-        transformer_block = _apply_ac_to_transformer_block(transformer_block, ac_config)
+        transformer_block = checkpoint_wrapper(transformer_block, ac_config)
         model.layers.register_module(layer_id, transformer_block)
 
     logger.info(f"Applied {ac_config.mode} activation checkpointing to the model")
+    return model
 
 
-def apply_compile(model: nn.Module):
-    # """
-    # Apply torch.compile to each TransformerBlock, which makes compilation efficient due to
-    # repeated structure. Alternatively one can compile the whole model (after applying DP).
-    # """
-    # for layer_id, transformer_block in model.layers.named_children():
-    #     transformer_block = torch.compile(transformer_block, fullgraph=True)
-    #     model.layers.register_module(layer_id, transformer_block)
+def apply_compile(model: nn.Module, job_config: JobConfig):
+    """Apply torch.compile to each transformer block."""
 
-    # logger.info("Compiling each TransformerBlock with torch.compile")
+    if job_config.model.norm_type == "fused_rmsnorm":
+        raise NotImplementedError(
+            "fused_rmsnorm is not compatible with torch.compile yet. Please use rmsnorm or layernorm."
+        )
+
+    # TODO(anijain): the following flag is on to accelarate compilation
+    #                remove it after it's enabled in pytorch by default
+    torch._dynamo.config.inline_inbuilt_nn_modules = True
 
     model = torch.compile(model)
     logger.info("Compiling the whole model with torch.compile")
@@ -335,21 +447,25 @@ def apply_compile(model: nn.Module):
     return model
 
 
-def apply_fsdp(
+def apply_dp(
     model: nn.Module,
-    dp_mesh: DeviceMesh,
-    param_dtype: torch.dtype,
-    reduce_dtype: torch.dtype,
-    pp_enabled: bool,
+    world_mesh: DeviceMesh,
+    parallel_dims: "ParallelDims",
+    job_config: JobConfig,
 ):
-    """
-    Apply data parallelism to the model. FSDP2 is used here.
-    """
-    mp_policy = MixedPrecisionPolicy(param_dtype=param_dtype, reduce_dtype=reduce_dtype)
+    """Apply data parallelism (FSDP2) to the model."""
+
+    dp_mesh = world_mesh["dp"] if world_mesh.ndim > 1 else world_mesh
+    assert dp_mesh.mesh_dim_names == ("dp",), dp_mesh.mesh_dim_names
+
+    mp_policy = MixedPrecisionPolicy(
+        param_dtype=TORCH_DTYPE_MAP[job_config.training.mixed_precision_param],
+        reduce_dtype=TORCH_DTYPE_MAP[job_config.training.mixed_precision_reduce],
+    )
     fsdp_config = {"mesh": dp_mesh, "mp_policy": mp_policy}
 
     for layer_id, transformer_block in model.layers.items():
-        if pp_enabled:
+        if parallel_dims.pp_enabled:
             # For PP, do not reshard after forward to avoid per-microbatch
             # all-gathers, which can be expensive and non-overlapped
             reshard_after_forward = False
@@ -362,37 +478,73 @@ def apply_fsdp(
             **fsdp_config,
             reshard_after_forward=reshard_after_forward,
         )
-    fully_shard(model, **fsdp_config, reshard_after_forward=not pp_enabled)
-
-    if pp_enabled:
-        # TODO
-        # This PR https://github.com/pytorch/pytorch/pull/129519 added a safety check to avoid using 2D/3D DCP since
-        # without strided sharding, DCP can not safely support resharding for 2D/3D.  However, for PP to work, even
-        # without resharding, we load a seed-checkpoint and need to disable the safety mechanism.  This hack should be
-        # removed after strided sharding is landed in DCP.
-        for module in model.modules():
-            assert len(module._load_state_dict_pre_hooks) <= 1
-            module._load_state_dict_pre_hooks.clear()
-            assert len(module._state_dict_pre_hooks) <= 1
-            module._state_dict_pre_hooks.clear()
+    fully_shard(
+        model, **fsdp_config, reshard_after_forward=not parallel_dims.pp_enabled
+    )
 
     logger.info("Applied FSDP to the model")
+    return model
 
 
-def apply_ddp(
+def parallelize_llama(
     model: nn.Module,
-    dp_mesh: DeviceMesh,
-    enable_compile: bool,
-    enable_compiled_autograd: bool,
+    world_mesh: DeviceMesh,
+    parallel_dims: "ParallelDims",
+    job_config: JobConfig,
 ):
-    if enable_compile:
-        if enable_compiled_autograd:
-            torch._dynamo.config.optimize_ddp = (
-                "python_reducer_without_compiled_forward"
-            )
-        else:
-            torch._dynamo.config.optimize_ddp = "ddp_optimizer"
+    """
+    Apply tensor parallelism, activation checkpointing, torch.compile, and data
+    parallelism to the model.
 
-    replicate(model, device_mesh=dp_mesh, bucket_cap_mb=100)
+    NOTE: The passed-in model preferably should be on meta device. Otherwise,
+    the model must fit on GPU or CPU memory.
+    """
+    # NOTE(lty): experimental for the PT-D 24 research internship project
+    if job_config.experimental.torch_spmd:
+        return parallelize_llama_torch_spmd(
+            model, world_mesh, parallel_dims, job_config
+        )
 
-    logger.info("Applied DDP to the model")
+    if parallel_dims.tp_enabled:
+        model = apply_tp(model, world_mesh, parallel_dims, job_config)
+
+    if job_config.activation_checkpoint.mode != "none":
+        model = apply_ac(model, job_config)
+
+    if parallel_dims.dp_enabled:
+        model = apply_dp(model, world_mesh, parallel_dims, job_config)
+
+    # whole-model compile should happen after FSDP wrapping
+    if job_config.training.compile:
+        model = apply_compile(model, job_config)
+
+    return model
+
+
+# NOTE(lty): experimental for the PT-D 24 research internship project
+def parallelize_llama_torch_spmd(
+    model, world_mesh, parallel_dims, job_config: JobConfig
+):
+    assert not parallel_dims.pp_enabled, "PP not supported by torch_spmd yet"
+    assert not parallel_dims.tp_enabled, "TP not supported by torch_spmd yet"
+    ac_config = job_config.activation_checkpoint
+    assert ac_config.mode == "none", "AC not supported by torch_spmd yet"
+
+    if parallel_dims.dp_enabled:
+        from torch_spmd.data_parallel import data_parallel, MixedPrecisionPolicy
+
+        mp_policy = MixedPrecisionPolicy(
+            param_dtype=TORCH_DTYPE_MAP[job_config.training.mixed_precision_param],
+            reduce_dtype=TORCH_DTYPE_MAP[job_config.training.mixed_precision_reduce],
+        )
+        dp_mesh = world_mesh["dp"] if world_mesh.ndim > 1 else world_mesh
+        model = data_parallel(model, dp_mesh, mode="fully_shard", mp_policy=mp_policy)
+        logger.info("Applied Simple FSDP to the model")
+
+    if job_config.training.compile:
+        # flag needed to make compile work on latest pytorch
+        torch._dynamo.config.inline_inbuilt_nn_modules = True
+        logger.info("Compiling with torch.compile")
+        model = torch.compile(model, fullgraph=True)
+
+    return model
