+++ torchtitan/parallelisms/pipelining_utils.py
@@ -7,28 +7,21 @@ from typing import Tuple
 
 from torch.distributed.pipelining import (
     Schedule1F1B,
-    ScheduleFlexibleInterleaved1F1B,
     ScheduleGPipe,
     ScheduleInterleaved1F1B,
 )
-from torchtitan.logging import logger
+from torchtitan.logging_utils import logger
 
 
-def build_pipeline_schedule(job_config, stages, loss_fn):
+def build_pipeline_schedule(job_config, parallel_dims, stages, loss_fn):
+
     looped_schedule = False
-
     if job_config.experimental.pipeline_parallel_schedule == "1f1b":
         schedule_class = Schedule1F1B
     elif job_config.experimental.pipeline_parallel_schedule == "gpipe":
         schedule_class = ScheduleGPipe
     elif job_config.experimental.pipeline_parallel_schedule == "interleaved_1f1b":
         schedule_class = ScheduleInterleaved1F1B
-        looped_schedule = True
-    elif (
-        job_config.experimental.pipeline_parallel_schedule
-        == "flexible_interleaved_1f1b"
-    ):
-        schedule_class = ScheduleFlexibleInterleaved1F1B
         looped_schedule = True
     else:
         raise NotImplementedError(
