+++ torchtitan/utils.py
@@ -4,7 +4,6 @@
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
 
-import gc
 import os
 from dataclasses import dataclass
 from datetime import timedelta
@@ -14,17 +13,18 @@ import torch
 import torch.distributed._functional_collectives as funcol
 import torch.distributed.distributed_c10d as c10d
 from torch.distributed.device_mesh import DeviceMesh
-from torchtitan.logging import logger
+from torchtitan.logging_utils import logger
+from torchtitan.parallelisms import ParallelDims
 
 
 def dist_max(x: Union[int, float], mesh: DeviceMesh) -> float:
     tensor = torch.tensor(x).cuda()
-    return funcol.all_reduce(tensor, reduceOp=c10d.ReduceOp.MAX.name, group=mesh).item()
+    return funcol.all_reduce(tensor, reduceOp=c10d.ReduceOp.MAX.name, group=mesh)
 
 
 def dist_mean(x: Union[int, float], mesh: DeviceMesh) -> float:
     tensor = torch.tensor(x).cuda()
-    return funcol.all_reduce(tensor, reduceOp=c10d.ReduceOp.AVG.name, group=mesh).item()
+    return funcol.all_reduce(tensor, reduceOp=c10d.ReduceOp.AVG.name, group=mesh)
 
 
 def _warn_overwrite_env(env, val):
@@ -33,6 +33,24 @@ def _warn_overwrite_env(env, val):
             f"ENV[{env}] = {os.environ[env]} will be overridden to {val} based on job config"
         )
     os.environ[env] = val
+
+
+def get_metrics_rank(world_mesh: DeviceMesh, parallel_dims: ParallelDims) -> int:
+    """
+    Returns global rank 0 in non-pipeline-parallel configs, and returns the global
+    rank of the 0th rank in the last pipeline stage when pipeline parallelism is enabled.
+    """
+    if parallel_dims.pp_enabled:
+        assert (
+            world_mesh.mesh_dim_names[0] == "pp"
+        ), "get_metrics_rank assumes pp is the outer mesh dim"
+        pp_mesh = world_mesh["pp"]
+        pp_size = pp_mesh.size()
+        metrics_log_rank = int((world_mesh.size() // pp_size) * (pp_size - 1))
+    else:
+        metrics_log_rank = 0
+
+    return metrics_log_rank
 
 
 def set_pg_timeouts(timeout, world_mesh):
@@ -60,19 +78,6 @@ def set_pg_timeouts(timeout, world_mesh):
     groups.append(None)
     for group in groups:
         torch.distributed.distributed_c10d._set_pg_timeout(timeout, group)
-
-
-# used to avoid stragglers in garbage collection
-class GarbageCollection:
-    def __init__(self, gc_freq=1000):
-        assert gc_freq > 0, "gc_freq must be a positive integer"
-        self.gc_freq = gc_freq
-        gc.disable()
-        gc.collect(1)
-
-    def run(self, step_count):
-        if step_count > 1 and step_count % self.gc_freq == 0:
-            gc.collect(1)
 
 
 TRACE_BUFFER_SIZE = "TORCH_NCCL_TRACE_BUFFER_SIZE"
