+++ train.py
@@ -5,43 +5,134 @@
 # LICENSE file in the root directory of this source tree.
 
 import contextlib
+import gc
 import os
 import time
+
+from dataclasses import dataclass, field
 from datetime import timedelta
+from io import BytesIO
+from timeit import default_timer as timer
+from typing import Any, Dict, List
+
+import numpy as np
 
 import torch
+import torch.nn.functional as F
+from torch.distributed import destroy_process_group
+from torch.distributed.checkpoint.stateful import Stateful
 from torch.distributed.elastic.multiprocessing.errors import record
+from torch.distributed.tensor.parallel import loss_parallel
 
-from torchtitan import utils
-from torchtitan.checkpoint import CheckpointManager, TrainState
+from torchtitan.checkpoint import CheckpointManager
 from torchtitan.config_manager import JobConfig
-from torchtitan.datasets import build_hf_data_loader, build_tokenizer
-from torchtitan.float8 import Float8Handler
-from torchtitan.logging import init_logger, logger
+from torchtitan.datasets import build_hf_data_loader, create_tokenizer
+from torchtitan.float8_linear import build_fp8_linear
+from torchtitan.logging_utils import init_logger, logger
+from torchtitan.lr_scheduling import get_lr_schedulers
 from torchtitan.metrics import build_gpu_memory_monitor, build_metric_logger
 from torchtitan.models import model_name_to_cls, model_name_to_tokenizer, models_config
-from torchtitan.optimizer import build_lr_schedulers, build_optimizers
 from torchtitan.parallelisms import (
     models_parallelize_fns,
     models_pipelining_fns,
     ParallelDims,
 )
+from torchtitan.parallelisms.pipelining_utils import build_pipeline_schedule
 from torchtitan.profiling import maybe_enable_memory_snapshot, maybe_enable_profiling
+from torchtitan.utils import (
+    Color,
+    dist_max,
+    dist_mean,
+    get_metrics_rank,
+    get_num_flop_per_token,
+    get_num_params,
+    get_peak_flops,
+    init_distributed,
+    NoColor,
+    set_pg_timeouts,
+)
 
 
-def get_train_context(enable_loss_parallel: bool, enable_compiled_autograd: bool):
-    @contextlib.contextmanager
-    def context():
-        with contextlib.ExitStack() as stack:
-            if enable_loss_parallel:
-                stack.enter_context(torch.distributed.tensor.parallel.loss_parallel())
-            if enable_compiled_autograd:
-                stack.enter_context(
-                    torch._dynamo.utils.maybe_enable_compiled_autograd(True)
-                )
-            yield
+@dataclass
+class TrainState(Stateful):
+    step: int = 0
+    global_avg_losses: List[float] = field(default_factory=list)
+    global_max_losses: List[float] = field(default_factory=list)
+    log_steps: List[int] = field(default_factory=list)
 
-    return context
+    def state_dict(self) -> Dict[str, Any]:
+        # Only checkpoint global_avg_losses and global_max_losses per log frequency
+        # to avoid sync overhead in every iteration.
+        global_avg_losses_bytes = BytesIO()
+        torch.save(self.global_avg_losses, global_avg_losses_bytes)
+        global_max_losses_bytes = BytesIO()
+        torch.save(self.global_max_losses, global_max_losses_bytes)
+        log_steps_bytes = BytesIO()
+        torch.save(self.log_steps, log_steps_bytes)
+        return {
+            "step": torch.tensor(self.step, dtype=torch.int32),
+            "global_avg_losses": global_avg_losses_bytes,
+            "global_max_losses": global_max_losses_bytes,
+            "log_steps": log_steps_bytes,
+        }
+
+    def load_state_dict(self, state_dict) -> None:
+        self.step = state_dict["step"].item()
+        state_dict["global_avg_losses"].seek(0)
+        self.global_avg_losses = torch.load(
+            state_dict["global_avg_losses"], weights_only=False
+        )
+        state_dict["global_max_losses"].seek(0)
+        self.global_max_losses = torch.load(
+            state_dict["global_max_losses"], weights_only=False
+        )
+        state_dict["log_steps"].seek(0)
+        self.log_steps = torch.load(state_dict["log_steps"], weights_only=False)
+
+
+def build_optimizers(model_parts, job_config: JobConfig):
+    """Wrap one optimizer per model part in an OptimizersContainer which provides a single
+    step() and zero_grad() method for all the child optimizers.
+    """
+
+    def _build_optimizer(model):
+        name = job_config.optimizer.name
+        lr = job_config.optimizer.lr
+        fused = job_config.optimizer.fused
+
+        # Common parameters for both optimizers
+        optimizer_kwargs = {
+            "lr": lr,
+            "betas": (0.9, 0.95),
+            "weight_decay": 0.1,
+            "fused": fused,
+            "foreach": not fused,
+        }
+        if name == "Adam":
+            # TODO: make the optimizer options configurable by toml/cmd args
+            optimizer = torch.optim.Adam(model.parameters(), **optimizer_kwargs)
+        elif name == "AdamW":
+            optimizer = torch.optim.AdamW(model.parameters(), **optimizer_kwargs)
+        else:
+            raise NotImplementedError(f"Optimizer {name} not added.")
+
+        return optimizer
+
+    class OptimizersContainer:
+        """Util for calling step/zero_grad on multiple optimizers needed for virtual pipeline stages"""
+
+        def __init__(self, optimizers):
+            self.optimizers = optimizers
+
+        def step(self):
+            for optimizer in self.optimizers:
+                optimizer.step()
+
+        def zero_grad(self):
+            for optimizer in self.optimizers:
+                optimizer.zero_grad()
+
+    return OptimizersContainer([_build_optimizer(model) for model in model_parts])
 
 
 # Enable debug tracing on failure: https://pytorch.org/docs/stable/elastic/errors.html
@@ -51,10 +142,12 @@ def main(job_config: JobConfig):
     logger.info(f"Starting job: {job_config.job.description}")
 
     # used for colorful printing
-    color = utils.Color if job_config.metrics.enable_color_printing else utils.NoColor
+    color = Color if job_config.metrics.enable_color_printing else NoColor
 
     # take control of garbage collection to avoid stragglers
-    gc_handler = utils.GarbageCollection(gc_freq=job_config.training.gc_freq)
+    _gc_freq = job_config.training.gc_freq
+    gc.disable()
+    gc.collect(1)
 
     # init distributed
     world_size = int(os.environ["WORLD_SIZE"])
@@ -64,20 +157,17 @@ def main(job_config: JobConfig):
         pp=job_config.experimental.pipeline_parallel_degree,
         world_size=world_size,
         enable_loss_parallel=job_config.training.enable_loss_parallel,
-        dp_type=job_config.training.data_parallel_type,
     )
     device = torch.device(f"cuda:{int(os.environ['LOCAL_RANK'])}")
     torch.cuda.set_device(device)
-    utils.init_distributed(job_config)
-    # initialize GPU memory monitor and get peak flops for MFU calculation
-    gpu_memory_monitor = build_gpu_memory_monitor()
-    gpu_peak_flops = utils.get_peak_flops(gpu_memory_monitor.device_name)
+    init_distributed(job_config)
 
     # build meshes
     world_mesh = parallel_dims.build_mesh(device_type="cuda")
     if parallel_dims.dp_enabled:
         dp_mesh = world_mesh["dp"]
-        dp_degree, dp_rank = dp_mesh.size(), dp_mesh.get_local_rank()
+        dp_degree = dp_mesh.size()
+        dp_rank = dp_mesh.get_local_rank()
     else:
         dp_degree, dp_rank = 1, 0
 
@@ -88,7 +178,7 @@ def main(job_config: JobConfig):
 
     # build tokenizer
     tokenizer_type = model_name_to_tokenizer[model_name]
-    tokenizer = build_tokenizer(tokenizer_type, job_config.model.tokenizer_path)
+    tokenizer = create_tokenizer(tokenizer_type, job_config.model.tokenizer_path)
 
     # build dataloader
     data_loader = build_hf_data_loader(
@@ -100,6 +190,15 @@ def main(job_config: JobConfig):
         dp_degree,
         dp_rank,
     )
+
+    # loss_parallel enables dispatching to efficient loss operators
+    loss_parallel_ctx = (
+        loss_parallel if parallel_dims.loss_parallel_enabled else contextlib.nullcontext
+    )
+
+    # loss fn can be shared by pipeline-parallel or non-pp execution
+    def loss_fn(pred, labels):
+        return F.cross_entropy(pred.flatten(0, 1), labels.flatten(0, 1))
 
     # build model (using meta init)
     model_cls = model_name_to_cls[model_name]
@@ -117,17 +216,16 @@ def main(job_config: JobConfig):
     # so we init on CPU, apply parallelism, and then move to GPU
     # with torch.device("meta"):
     with torch.device("cpu"):
-        model = model_cls.from_model_args(model_config)
+        whole_model = model_cls.from_model_args(model_config)
 
-    # a no-op hander if float8 is not enabled
-    float8_handler = Float8Handler(job_config, parallel_dims)
-    # swap to Float8Linear based on float8 configs
-    float8_handler.convert_to_float8_training(model)
+    # apply fp8 linear module swap
+    if job_config.training.fp8_linear:
+        build_fp8_linear(whole_model, job_config)
 
     # log model size
-    model_param_count = utils.get_num_params(model)
-    num_flop_per_token = utils.get_num_flop_per_token(
-        utils.get_num_params(model, exclude_embedding=True),
+    model_param_count = get_num_params(whole_model)
+    num_flop_per_token = get_num_flop_per_token(
+        get_num_params(whole_model, exclude_embedding=True),
         model_config,
         job_config.training.seq_len,
     )
@@ -136,49 +234,43 @@ def main(job_config: JobConfig):
         f"{color.red}size: {model_param_count:,} total parameters{color.reset}"
     )
 
-    # loss function to be shared by Pipeline Parallel and SPMD training
-    def loss_fn(pred, labels):
-        return torch.nn.functional.cross_entropy(
-            pred.flatten(0, 1), labels.flatten(0, 1)
-        )
+    # initialize GPU memory monitor before applying parallelisms to the model
+    gpu_memory_monitor = build_gpu_memory_monitor()
+    # obtain the peak flops of bf16 type for MFU calculation
+    gpu_peak_flops = get_peak_flops(gpu_memory_monitor.device_name)
 
-    # apply parallelisms and initialization
     if parallel_dims.pp_enabled:
-        # apply PT-D Pipeline Parallel
-        pp_schedule, model_parts = models_pipelining_fns[model_name](
-            model, pp_mesh, parallel_dims, job_config, device, model_config, loss_fn
+        stages, model_parts = models_pipelining_fns[model_name](
+            whole_model, world_mesh, parallel_dims, job_config, device, model_config
         )
-
-        # For PP with looped schedules, each item in model_parts is one stage-model-chunk.
-        # We need to iterate through model_parts to apply SPMD parallelisms, compilation,
-        # optimizer, and checkpointing
-        for m in model_parts:
-            # apply SPMD-style PT-D techniques
-            models_parallelize_fns[model_name](m, world_mesh, parallel_dims, job_config)
-
-            # In PP, we cannot call init_weights directly because some layers are missing.
-            # In the future, we may make init_weights handle missing layers, but also have
-            # to consider RNG seed propagation. For now, we rely on a seed checkpoint to
-            # initialize the model.
-            m.to_empty(device="cuda")
-            m.train()
     else:
-        # apply PT-D Tensor Parallel, activation checkpointing, torch.compile, Data Parallel
-        model = models_parallelize_fns[model_name](
-            model, world_mesh, parallel_dims, job_config
-        )
+        # In 1D/2D cases or PP with simple schedules, model_parts is just one item
+        # for PP with looped schedules, each item is one stage-model-chunk
+        # we iterate all model_parts for applying SPMD parallelism, compilation, optimizer, and checkpointing
+        model_parts = [whole_model]
 
-        # move sharded model to CPU/GPU and initialize weights via DTensor
-        # init_device = "cpu" if job_config.checkpoint.create_seed_checkpoint else "cuda"
-        # model.to_empty(device=init_device)
-        # model.init_weights()
+    # apply PT-D DP/TP parallelisms and activation checkpointing
+    model_parts = [
+        models_parallelize_fns[model_name](m, world_mesh, parallel_dims, job_config)
+        for m in model_parts
+    ]
 
-        # NOTE(lty): move from CPU to GPU after applying FSDP2/torch_spmd parallelisms
-        # The original code related to meta init is commented out above
+    # NOTE(lty): move from CPU to GPU after applying FSDP2/torch_spmd parallelisms
+    # The original code related to meta init is commented out below
+    for model in model_parts:
         model.to("cuda")
-        model.train()
 
-        model_parts = [model]
+    # init_device = "cpu" if job_config.checkpoint.create_seed_checkpoint else "cuda"
+    # for model in model_parts:
+    #     model.to_empty(device=init_device)
+
+    if parallel_dims.pp_enabled:
+        pp_schedule = build_pipeline_schedule(job_config, parallel_dims, stage, loss_fn)
+    # else:
+    #     # If PP is enabled, we can't rely on init_weights, because some layers are missing.
+    #     # In the future, we may make init_weights handle missing layers, but also have to consider RNG seed propagation.
+    #     # allocate sharded model on GPU and initialize weights via DTensor
+    #     whole_model.init_weights()
 
     gpu_mem_stats = gpu_memory_monitor.get_peak_stats()
     logger.info(
@@ -189,16 +281,24 @@ def main(job_config: JobConfig):
 
     # build optimizer after applying parallelisms to the model
     optimizers = build_optimizers(model_parts, job_config)
-    lr_schedulers = build_lr_schedulers(optimizers.optimizers, job_config)
+    lr_schedulers = get_lr_schedulers(optimizers.optimizers, job_config)
+
+    metric_logger = build_metric_logger(
+        job_config, metrics_log_rank=get_metrics_rank(world_mesh, parallel_dims)
+    )
 
     train_state = TrainState()
 
+    # train loop
+    for model in model_parts:
+        model.train()
+
     # load initial checkpoint
     checkpoint = CheckpointManager(
-        dataloader=data_loader,
         model_parts=model_parts,
         optimizers=optimizers.optimizers,
         lr_schedulers=lr_schedulers.schedulers,
+        dataloader=data_loader,
         states={"train_state": train_state},
         job_config=job_config,
     )
@@ -219,8 +319,6 @@ def main(job_config: JobConfig):
             "Please run `./create_seed_checkpoint.sh` and rerun training with `--checkpoint.enable_checkpoint`"
         )
 
-    metric_logger = build_metric_logger(job_config, parallel_dims)
-
     # plot losses loaded from checkpoint (if any) to TensorBoard
     # NOTE: Loss info after the last log step before checkpoint saving will not be ploted.
     #       This can be avoided by setting checkpoint.interval to be a multiple of metrics.log_freq
@@ -234,29 +332,17 @@ def main(job_config: JobConfig):
 
     data_iterator = iter(data_loader)
 
-    train_context = get_train_context(
-        parallel_dims.loss_parallel_enabled,
-        job_config.experimental.enable_compiled_autograd,
-    )
-
-    # variables used to keep info for metrics logging
-    losses_since_last_log = []
-    ntokens_since_last_log = 0
-    data_loading_times = []
-    time_last_log = time.perf_counter()
-    gpu_memory_monitor.reset_peak_stats()
-
     checkpoint.reset()
 
+    # variables used to keep info for metrics logging
+    losses_since_last_log: List[float] = []
+    ntokens_since_last_log = 0
+    data_loading_times: List[float] = []
+    time_last_log = timer()
+    gpu_memory_monitor.reset_peak_stats()
+
     # train loop
-    logger.info(
-        f"Training starts at step {train_state.step + 1}, "
-        f"with local batch size {job_config.training.batch_size}, "
-        f"global batch size {job_config.training.batch_size * dp_degree}, "
-        f"sequence length {job_config.training.seq_len}, "
-        f"total steps {job_config.training.steps} "
-        f"(warmup {job_config.training.warmup_steps})"
-    )
+    logger.info(f"Training starts at step {train_state.step + 1}")
     with maybe_enable_profiling(
         job_config, global_step=train_state.step
     ) as torch_profiler, maybe_enable_memory_snapshot(
@@ -264,24 +350,25 @@ def main(job_config: JobConfig):
     ) as memory_profiler:
         while train_state.step < job_config.training.steps:
             train_state.step += 1
-            gc_handler.run(train_state.step)
+            if train_state.step > 1 and train_state.step % _gc_freq == 0:
+                gc.collect(1)
 
             # get batch
-            data_load_start = time.perf_counter()
+            data_load_start = timer()
             batch = next(data_iterator)
             input_ids, labels = batch
             ntokens_since_last_log += labels.numel()
-            data_loading_times.append(time.perf_counter() - data_load_start)
+            data_loading_times.append(timer() - data_load_start)
 
             input_ids = input_ids.cuda()
             labels = labels.cuda()
             optimizers.zero_grad()
 
             if parallel_dims.pp_enabled:
-                # Pipeline Parallel forward / backward inside step() call
+                # pipeline parallel forward / backward inside step() call
                 is_last_stage = pp_mesh.get_local_rank() == pp_mesh.size() - 1
 
-                with train_context():
+                with loss_parallel_ctx():
                     if pp_mesh.get_local_rank() == 0:
                         pp_schedule.step(input_ids)
                     elif is_last_stage:
@@ -298,7 +385,7 @@ def main(job_config: JobConfig):
                 )
             else:
                 # Non-PP forward / backward
-                with train_context():
+                with loss_parallel_ctx():
                     pred = model(input_ids)
                     loss = loss_fn(pred, labels)
                     # pred.shape=(bs, seq_len, vocab_size)
@@ -307,22 +394,15 @@ def main(job_config: JobConfig):
                     loss.backward()
 
             # clip gradients
-            for m in model_parts:
+            for model in model_parts:
                 torch.nn.utils.clip_grad_norm_(
-                    m.parameters(), job_config.training.max_norm, foreach=True
+                    model.parameters(), job_config.training.max_norm, foreach=True
                 )
 
-            # sync float8 amaxes and scales
-            float8_handler.sync_float8_amax_and_scale_history(model_parts)
-
             # optimizer step
-            checkpoint.maybe_wait_for_staging()
+            checkpoint.wait_for_staging()
             optimizers.step()
             lr_schedulers.step()
-
-            # calculate float8 dynamic amax/scale for all-parameter for FSDP2
-            # it issues a single all-reduce for all parameters at once for better performance
-            float8_handler.precompute_float8_dynamic_scale_for_fsdp(model_parts)
 
             losses_since_last_log.append(loss)
 
@@ -332,21 +412,23 @@ def main(job_config: JobConfig):
                 or train_state.step % job_config.metrics.log_freq == 0
             ):
                 losses = [loss.item() for loss in losses_since_last_log]
-                avg_loss, max_loss = sum(losses) / len(losses), max(losses)
+                avg_loss, max_loss = (
+                    np.mean(losses),
+                    np.max(losses),
+                )
                 if parallel_dims.dp_enabled:
                     global_avg_loss, global_max_loss = (
-                        utils.dist_mean(avg_loss, dp_mesh),
-                        utils.dist_max(max_loss, dp_mesh),
+                        dist_mean(avg_loss, dp_mesh).item(),
+                        dist_max(max_loss, dp_mesh).item(),
                     )
                 else:
                     global_avg_loss, global_max_loss = avg_loss, max_loss
 
-                # update train state
                 train_state.log_steps.append(train_state.step)
                 train_state.global_avg_losses.append(global_avg_loss)
                 train_state.global_max_losses.append(global_max_loss)
 
-                time_delta = time.perf_counter() - time_last_log
+                time_delta = timer() - time_last_log
 
                 # tokens per second, abbr. as wps by convention
                 wps = ntokens_since_last_log / (
@@ -358,8 +440,8 @@ def main(job_config: JobConfig):
                 mfu = 100 * num_flop_per_token * wps / gpu_peak_flops
 
                 time_end_to_end = time_delta / job_config.metrics.log_freq
-                time_data_loading = sum(data_loading_times) / len(data_loading_times)
-                time_data_loading_pct = 100 * sum(data_loading_times) / time_delta
+                time_data_loading = np.mean(data_loading_times)
+                time_data_loading_pct = 100 * np.sum(data_loading_times) / time_delta
 
                 gpu_mem_stats = gpu_memory_monitor.get_peak_stats()
 
@@ -392,23 +474,23 @@ def main(job_config: JobConfig):
                 losses_since_last_log.clear()
                 ntokens_since_last_log = 0
                 data_loading_times.clear()
-                time_last_log = time.perf_counter()
+                time_last_log = timer()
                 gpu_memory_monitor.reset_peak_stats()
 
             checkpoint.save(
                 train_state.step, force=(train_state.step == job_config.training.steps)
             )
 
-            # signal the profiler that the next profiling step has started
+            # signals the profiler that the next profiling step has started
             if torch_profiler:
                 torch_profiler.step()
+
             if memory_profiler:
                 memory_profiler.step()
 
-            # reduce timeout after first train step for faster signal
-            # (assuming lazy init and compilation are finished)
+            # Reduce timeout after first train step for faster signal (assumes lazy init, compile are finished)
             if train_state.step == 1:
-                utils.set_pg_timeouts(
+                set_pg_timeouts(
                     timeout=timedelta(seconds=job_config.comm.train_timeout_seconds),
                     world_mesh=world_mesh,
                 )
@@ -425,4 +507,4 @@ if __name__ == "__main__":
     config = JobConfig()
     config.parse_args()
     main(config)
-    torch.distributed.destroy_process_group()
+    destroy_process_group()
